Split save_stages.rs (874 lines) into per-domain submodules
Break the monolithic save_stages.rs into 7 focused files under
save_stages/ directory, one per pipeline stage plus assembly:

- grid_stage.rs (78 lines) - GridStageOutput + collect_grid_stage
- economy_stage.rs (79 lines) - EconomyStageOutput + collect_economy_stage
- entity_stage.rs (149 lines) - EntityStageOutput + collect_entity_stage
- environment_stage.rs (122 lines) - EnvironmentStageOutput + collect_environment_stage
- disaster_stage.rs (242 lines) - DisasterStageOutput + collect_disaster_stage
- policy_stage.rs (117 lines) - PolicyStageOutput + collect_policy_stage
- assemble.rs (70 lines) - assemble_save_data

All pub items re-exported from mod.rs so callers are unaffected.

Part of #1389
